backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
google-generativeai==0.8.3
openai>=2.0.0
pydub==0.25.1
audioop-lts==0.2.1
python-dotenv==1.0.1
